# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Convoy API",
    "description": "\nConvoy is a truck discovery and availability marketplace API.\n\n## Authentication\n1. **POST /api/auth/send-otp** then **POST /api/auth/verify-otp** to log in.\n2. Use the returned **`accessToken`** as a Bearer JWT on all other endpoints:\n   `Authorization: Bearer <accessToken>`\n3. Path `user_id` must match the logged-in user (from the token).\n\n## Interactive docs\n- **Swagger UI:** `/docs`\n- **ReDoc:** `/redoc`\n- **OpenAPI JSON:** `/openapi.json`\n\n## Base path\nAll routes are served under **`/api`**.\n",
    "contact": {
      "name": "Convoy Support",
      "url": "https://github.com/SkandTrizip/convoy_prod"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Current environment"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/search/trucks": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Trucks",
        "description": "Search for available trucks using PostGIS ST_DWithin.",
        "operationId": "search_trucks_api_search_trucks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchTrucksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/search/track-demand/{user_id}": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Track Search Demand",
        "description": "Track failed search for smart notifications",
        "operationId": "track_search_demand_api_search_track_demand__user_id__post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchTrucksRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/log-call/{user_id}": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Log Call Click",
        "description": "Log call button click for analytics",
        "operationId": "log_call_click_api_search_log_call__user_id__post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Location": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable place name"
          },
          "lat": {
            "type": "number",
            "title": "Lat",
            "description": "Latitude"
          },
          "lng": {
            "type": "number",
            "title": "Lng",
            "description": "Longitude"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Location source type",
            "default": "local"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode",
            "description": "Postal code"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City",
            "description": "City name"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "State name"
          }
        },
        "type": "object",
        "required": [
          "name",
          "lat",
          "lng"
        ],
        "title": "Location",
        "example": {
          "city": "Mumbai",
          "lat": 19.076,
          "lng": 72.8777,
          "name": "Mumbai, Maharashtra",
          "pincode": "400001",
          "state": "Maharashtra",
          "type": "local"
        }
      },
...
      "SearchTrucksRequest": {
        "properties": {
          "origin": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required if destination is omitted \u2014 at least one of the two is needed"
          },
          "destination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required if origin is omitted \u2014 at least one of the two is needed"
          },
          "truckType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trucktype",
            "description": "Truck type filter. If omitted, posts of any truck type are returned.",
            "examples": [
              "Open Body"
            ]
          },
          "radius_km": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Radius Km",
            "description": "Search radius in kilometres",
            "default": 150
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "Page number (1-indexed), 10 results per page",
            "default": 1
          },
          "capacity": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity",
            "description": "Minimum required capacity in tonnes. If omitted, posts of any capacity are returned; if set, only posts with capacity >= this value are returned."
          },
          "sortBy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "nearest_origin",
                  "nearest_destination"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sortby",
            "description": "Sort results by distance to origin or destination. If omitted, defaults to combined distance (both given), or whichever single one was given. If the requested side wasn't provided in the search (e.g. nearest_origin with no origin), falls back to the default ranking instead of failing."
          }
        },
        "type": "object",
        "title": "SearchTrucksRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /api/auth/verify-otp (`accessToken`)"
      }
    }
  },
...
}